Major improvements on resource usage monitoring
- Remove unused events
- Move event listener to separate class
- Handle FailureEvent to reconnect a websocket connection in case it disconnects
- Implemented a potential fix for websocket-based resource monitor not stopping upon calling ResourceUsageManager#stopListener()
