+++ src/main/java/com/sparkedhost/accuratereadings/managers/ResourceUsageManager.java
@@ -7,9 +7,11 @@ import com.mattmalec.pterodactyl4j.client.managers.WebSocketManager;
 import com.mattmalec.pterodactyl4j.client.ws.events.AuthSuccessEvent;
 import com.mattmalec.pterodactyl4j.client.ws.events.StatsUpdateEvent;
 import com.mattmalec.pterodactyl4j.client.ws.events.connection.DisconnectedEvent;
+import com.mattmalec.pterodactyl4j.client.ws.events.connection.FailureEvent;
 import com.mattmalec.pterodactyl4j.client.ws.events.token.TokenEvent;
 import com.mattmalec.pterodactyl4j.client.ws.hooks.ClientSocketListenerAdapter;
 import com.sparkedhost.accuratereadings.Main;
+import com.sparkedhost.accuratereadings.events.WebSocketEvents;
 import lombok.Getter;
 import lombok.Setter;
 import org.bukkit.Bukkit;
@@ -19,7 +21,6 @@ import java.util.logging.Level;
 
 public class ResourceUsageManager extends ClientSocketListenerAdapter {
     final PterodactylManager pteroManager = Main.getInstance().pteroAPI;
-    private ResourceUsageManager resManagerThread;
 
     @Getter
     private BukkitTask fallbackTimer;
@@ -27,6 +28,10 @@ public class ResourceUsageManager extends ClientSocketListenerAdapter {
     @Getter
     @Setter
     private boolean isRunning = false;
+
+    @Getter
+    @Setter
+    private WebSocketManager webSocketManager;
 
     /**
      * Starts resource usage listener.
@@ -38,9 +43,8 @@ public class ResourceUsageManager extends ClientSocketListenerAdapter {
 
         // If use-websocket is set to true in the config, use that to gather resource usage stats.
         if (Main.getInstance().getSettings().pterodactyl_useWebsocket) {
-            resManagerThread = new ResourceUsageManager();
             pteroManager.getApi().retrieveServerByIdentifier(pteroManager.getServerId()).map(ClientServer::getWebSocketBuilder)
-                    .map(builder -> builder.addEventListeners(resManagerThread)).executeAsync(WebSocketBuilder::build);
+                    .map(builder -> builder.addEventListeners(new WebSocketEvents())).executeAsync(WebSocketBuilder::build);
             return;
         }
 
@@ -60,46 +64,15 @@ public class ResourceUsageManager extends ClientSocketListenerAdapter {
      */
     public void stopListener() {
         setRunning(false);
-        Main.getInstance().log(Level.INFO, "Resource usage monitor has been stopped.");
 
-        if (getFallbackTimer() == null) {
-            pteroManager.getApi().retrieveServerByIdentifier(pteroManager.getServerId()).map(ClientServer::getWebSocketBuilder)
-                    .map(builder -> builder.removeEventListeners(resManagerThread)).executeAsync();
-            return;
+        if (getFallbackTimer() == null && getWebSocketManager() != null) {
+            getWebSocketManager().shutdown();
+            setWebSocketManager(null);
+        } else {
+            getFallbackTimer().cancel();
         }
 
-        getFallbackTimer().cancel();
         pteroManager.resetVariables();
-    }
-
-    /*
-     * The following event listeners are only registered when using a websocket connection to poll resource usage
-     * statistics, and remain unused when using the standard API polling method.
-     */
-
-    @Override
-    public void onAuthSuccess(AuthSuccessEvent e) {
-        Main.getInstance().log(Level.INFO, "Successfully established a websocket connection.");
-        e.getWebSocketManager().request(WebSocketManager.RequestAction.STATS);
-    }
-
-    @Override
-    public void onStatsUpdate(StatsUpdateEvent e) {
-        pteroManager.setCpuUsage((long) e.getCPU());
-        pteroManager.setMemoryUsage(e.getMemory());
-        pteroManager.setDiskUsage(e.getDisk());
-        pteroManager.setUptime(e.getUptimeFormatted());
-    }
-
-    @Override
-    public void onTokenUpdate(TokenEvent e) {
-        Main.getInstance().log(Level.INFO, "The authentication token has been updated.");
-    }
-
-    @Override
-    public void onDisconnected(DisconnectedEvent e) {
-        Main.getInstance().log(Level.WARNING, "Websocket connection lost, restarting resource usage manager...");
-        stopListener();
-        Bukkit.getScheduler().runTaskLater(Main.getInstance(), this::startListener, 60L);
+        Main.getInstance().log(Level.INFO, "Resource usage monitor has been stopped.");
     }
 }
